[Refactor] deploment target에 맞게 UI 수정
